feat: got precision 16 back
Former-commit-id: 6b19dc9bd17078bb2c151d5cd96e7ba4da9e1b89 [formerly 5d1eac2ed10be960c89407ad265ff350e11c1adf]
Former-commit-id: 1db4ca0ce11ac818408b94625b872c1202b5d4ed

+++ src/unet/module.py
@@ -38,7 +38,7 @@ class UNetModule(pl.LightningModule):
 
         # forward pass, compute masks
         prediction = self.model(data)
-        binary = (torch.sigmoid(prediction) > 0.5).float()  # TODO: check if float necessary
+        binary = (torch.sigmoid(prediction) > 0.5).half()
 
         # compute metrics (in dictionnary)
         metrics = {
